refactor(StudentDataService): Extract CompletionService
