+++ src/app/services/data.service.ts
@@ -15,8 +15,6 @@ export class DataService {
 
   constructor(protected ProjectService: ProjectService) {}
 
-  isCompleted(nodeId, componentId) {}
-
   endCurrentNodeAndSetCurrentNodeByNodeId(nextNodeId) {}
 
   getCurrentNode() {
