+++ src/assets/wise5/common/constraint/strategies/IsCompletedConstraintStrategy.ts
@@ -2,6 +2,6 @@ import { AbstractConstraintStrategy } from './AbstractConstraintStrategy';
 
 export class IsCompletedConstraintStrategy extends AbstractConstraintStrategy {
   evaluate(criteria: any): boolean {
-    return this.dataService.isCompleted(criteria.params.nodeId);
+    return this.completionService.isCompleted(criteria.params.nodeId);
   }
 }
